♻️ refactor: Changed service logic to RepositoryImpl.cs, reducing dependency on service classes.
