+++ NovaModaBrecho/Services/ItemService.cs
@@ -1,42 +1,23 @@
 using NovaModaBrecho.Models;
+using NovaModaBrecho.Repository.Interfaces;
 using NovaModaBrecho.Services.Interfaces;
 
 namespace NovaModaBrecho.Services;
 
-public class ItemService<T> : IBaseItemService<T> where T : Item
+public class ItemService<T>(IRepository<T> repository) : IBaseItemService<T>
+    where T : Item
 {
-    private readonly List<T> _items = [];
-    private int _nextId = 1;
-    public List<T> GetAll()
-    {
-        return _items;
-    }
+    public List<T> GetAll() => repository.GetAll();
 
-    public T? GetById(int id)
-    {
-        return _items.FirstOrDefault(x => x.Id == id);
-    }
+    public T? GetById(int id) => repository.GetById(id);
 
     public void Add(T? item)
     {
         if (item != null)
-        {
-            item.Id = ++_nextId;
-            _items.Add(item);
-        }
+            repository.Add(item);
     }
 
-    public void Update(T item)
-    {
-        var index = _items.FindIndex(i => i.Id == item.Id);
-        if (index != -1)
-        {
-            _items[index] = item;
-        }
-    }
+    public void Update(T item) => repository.Update(item);
 
-    public void Delete(int id)
-    {
-        _items.RemoveAll(x => x.Id == id);
-    }
+    public void Delete(int id) => repository.Delete(id);
 }
